refactor: inline dex routes
