--- apps/main/src/app/dex/[network]/compensation/page.tsx
@@ -1,8 +1,0 @@
-import { PageCompensation } from '@/dex/components/PageCompensation/Page'
-import type { NetworkUrlParams } from '@/dex/types/main.types'
-import { useParams } from '@ui-kit/hooks/router'
-
-export default function Component() {
-  const params = useParams<NetworkUrlParams>()
-  return <PageCompensation {...params} />
-}
