--- apps/main/src/app/dex/[network]/create-pool/page.tsx
@@ -1,8 +1,0 @@
-import { PageCreatePool } from '@/dex/components/PageCreatePool/Page'
-import type { NetworkUrlParams } from '@/dex/types/main.types'
-import { useParams } from '@ui-kit/hooks/router'
-
-export default function Component() {
-  const params = useParams<NetworkUrlParams>()
-  return <PageCreatePool {...params} />
-}
