--- apps/main/src/app/dex/[network]/disclaimer/page.tsx
@@ -1,5 +1,0 @@
-import { Disclaimer } from '@ui-kit/widgets/Disclaimer'
-
-export default function Component() {
-  return <Disclaimer currentApp="dex" />
-}
