--- apps/main/src/app/dex/[network]/integrations/page.tsx
@@ -1,8 +1,0 @@
-import Integrations from '@/dex/components/PageIntegrations/Page'
-import type { NetworkUrlParams } from '@/dex/types/main.types'
-import { useParams } from '@ui-kit/hooks/router'
-
-export default function Component() {
-  const params = useParams<NetworkUrlParams>()
-  return <Integrations {...params} />
-}
