--- apps/main/src/app/dex/[network]/pools/page.tsx
@@ -1,8 +1,0 @@
-import { PagePoolList } from '@/dex/components/PagePoolList/Page'
-import type { NetworkUrlParams } from '@/dex/types/main.types'
-import { useParams } from '@ui-kit/hooks/router'
-
-export default function Component() {
-  const params = useParams<NetworkUrlParams>()
-  return <PagePoolList {...params} />
-}
