+++ apps/main/src/dex/components/PageCompensation/Page.tsx
@@ -13,9 +13,11 @@ import IconButton from '@ui/IconButton'
 import ExternalLink from '@ui/Link/ExternalLink'
 import Spinner, { SpinnerWrapper } from '@ui/Spinner'
 import { isLoading, useConnection, useWallet } from '@ui-kit/features/connect-wallet'
+import { useParams } from '@ui-kit/hooks/router'
 import { t } from '@ui-kit/lib/i18n'
 
-export const PageCompensation = ({ network }: NetworkUrlParams) => {
+export const PageCompensation = () => {
+  const { network } = useParams<NetworkUrlParams>()
   const { curveApi = null, connectState } = useConnection()
   const { connect: connectWallet, provider } = useWallet()
   const [contracts, setContracts] = useState<EtherContract[]>([])
