+++ apps/main/src/dex/components/PagePoolList/Page.tsx
@@ -12,7 +12,7 @@ import type { NetworkUrlParams } from '@/dex/types/main.types'
 import { getPath } from '@/dex/utils/utilsRouter'
 import { breakpoints } from '@ui/utils/responsive'
 import { useConnection } from '@ui-kit/features/connect-wallet'
-import { useNavigate, useSearchParams } from '@ui-kit/hooks/router'
+import { useNavigate, useSearchParams, useParams } from '@ui-kit/hooks/router'
 import { t } from '@ui-kit/lib/i18n'
 
 enum SEARCH {
@@ -24,7 +24,8 @@ enum SEARCH {
 
 type PageProps = NetworkUrlParams
 
-export const PagePoolList = (params: PageProps) => {
+export const PagePoolList = () => {
+  const params = useParams<PageProps>()
   const push = useNavigate()
   const searchParams = useSearchParams()
   const { curveApi = null } = useConnection()
